refactor: rename package and rearrange structure
